Remove About and add content to Connect
