+++ src/locales/en.json
@@ -1,9 +1,10 @@
 {
   "_display_name": "English",
-  "about": {
-    "description": "The {{applicationName}} app is made available by {{healthAuthorityName}}",
+  "connect": {
+    "description": "The {{applicationName}} app is made available by {{healthAuthorityName}}.\n\nFor more information, visit the ",
     "operating_system_abbr": "OS:",
-    "version": "Version:"
+    "version": "Version:",
+    "ha_link": "{{healthAuthorityName}} website."
   },
   "assessment": {
     "agree_question_description": "Stop and call emergency services if you are experiencing:\n• Severe, constant chest pain or pressure \n• Extreme difficulty breathing \n• Severe, constant lightheadedness \n• Serious disorientation or unresponsiveness",
